Fix code for 2021 version of BBC site
The pages were redesigned.

Retrieve a smaller subset of information from the category pages only.

One limitation is that this only includes the most recent episode where
multiple episodes are grouped together. However, it reduces the number
of page requests: we only process the category pages, not individual
pages.

Also run Black to tidy the code.